Title: 두 실수의 곱, Time: 5ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240404/두 실수의 곱/the-product-of-two-real-numbers.c
@@ -2,8 +2,8 @@
 
 int main() {
     // 여기에 코드를 작성해주세요.
-    int a = 5.26;
-    int b = 8.27;
+    float a = 5.26;
+    float b = 8.27;
 
     printf("%.3f", a*b);
     return 0;
